+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [[#232]](https://github.com/Marcin-Migdal/m-component-library/issues/232) **[PATCH]** Add Responsive behavior to the Breadcrumb when it does not fit in its container
 - [[#249]](https://github.com/Marcin-Migdal/m-component-library/issues/249) **[PATCH]** DropdownMenu position calculation on window resize
 - [[#246]](https://github.com/Marcin-Migdal/m-component-library/issues/246) **[PATCH]** Fix StandAloneTextfield component styling when passing children props to the component
+- [[#247]](https://github.com/Marcin-Migdal/m-component-library/issues/247) **[PATCH]** Delete unsupported prop from Textarea input props object
 
 ### Release date 04.07.2025 - Release 2.1.5